nâng cấp chat bot, sửa vài lỗi bên spring

+++ SpringBoot-CuaHangDienThoai/src/main/java/com/cuahangdienthoai/service/RecommendationService.java
@@ -2,7 +2,7 @@ package com.cuahangdienthoai.service;
 
 public interface RecommendationService {
     String getRecommendOfDevices(Long productID);
-    String getRecommendOfUser(int userId);
+    String getRecommendOfUser(Long userId);
 
     void train();
 }
